PR Code Cleanup: Use Long instead of Integer for countbreedingmethods Create countFilteredMethods in MethodDAO IBP-4252-BrapiBreedingMethods

+++ src/main/java/org/generationcp/middleware/api/breedingmethod/BreedingMethodService.java
@@ -11,5 +11,5 @@ public interface BreedingMethodService {
 
 	List<BreedingMethodDTO> getBreedingMethods(BreedingMethodSearchRequest methodSearchRequest);
 
-	Integer countBreedingMethods(BreedingMethodSearchRequest methodSearchRequest);
+	Long countBreedingMethods(BreedingMethodSearchRequest methodSearchRequest);
 }
